Turned macros unit tests into integration tests
